Remove shard relay subscription for init config
Being able to set what shard to subscribe too when initialising a Waku node is a biased towards service nodes.

Indeed, subscribing to shard should be done with the `subscribe` API (yet to be defined). Which can be used with a content topic, and pubsub topic (only if static sharding).

A service node such as the `wakunode2` binary can still accept "shards to subscribe to" via CLI, but this is unrelated to the Waku node conf, and the `wakunode2` software should simply call `subscribe` when ready.